+++ src/alcotest-engine/core.ml
@@ -15,6 +15,7 @@
  *)
 
 open Astring
+open Model
 open Utils
 
 let compare_int : int -> int -> int = compare
@@ -22,12 +23,10 @@ let compare_int : int -> int -> int = compare
 module IntSet = Set.Make (struct
   type t = int
 
-  let compare : int -> int -> int = compare
+  let compare = compare_int
 end)
 
-exception Registration_error of string
-
-exception Check_error of string
+exception Check_error of unit Fmt.t
 
 module type S = sig
   type return
@@ -68,7 +67,22 @@ module Make (P : Platform.MAKER) (M : Monad.S) : S with type return = unit M.t =
 struct
   module P = P (M)
   module M = Monad.Extend (M)
+  module Suite = Suite (M)
   include M.Infix
+
+  (** Take a string path and collapse a leading [$HOME] path segment to [~]. *)
+  let maybe_collapse_home path =
+    match P.home_directory () with
+    | Error _ -> path
+    | Ok home -> (
+        (* Astring doesn't have [cut_prefix]. *)
+        match String.is_prefix ~affix:home path with
+        | false -> path
+        | true ->
+            let tail =
+              String.Sub.to_string (String.sub ~start:(String.length home) path)
+            in
+            "~" ^ tail )
 
   (* Types *)
   type return = unit M.t
@@ -79,87 +93,18 @@ struct
 
   exception Test_error
 
-  let compare_path (`Path (s, i)) (`Path (s', i')) =
-    match String.compare s s' with 0 -> compare_int i i' | n -> n
-
-  let short_string_of_path (`Path (n, i)) = Printf.sprintf "%s.%03d" n i
-
-  let file_of_path (`Path (n, i)) =
-    let path = `Path (String.Ascii.lowercase n, i) in
-    Printf.sprintf "%s.output" (short_string_of_path path)
-
-  type 'a rrun = 'a -> Pp.run_result M.t
-
   type 'a test_case = string * speed_level * 'a run
 
   let test_case n s f = (n, s, f)
 
   type 'a test = string * 'a test_case list
 
-  module Suite : sig
-    type 'a t
-
-    type 'a test_case = {
-      path : Pp.path;
-      speed_level : speed_level;
-      fn : 'a rrun;
-    }
-
-    val empty : unit -> 'a t
-
-    val add : 'a t -> Pp.path * string * speed_level * 'a rrun -> 'a t
-
-    val tests : 'a t -> 'a test_case list
-
-    val doc_of_path : 'a t -> Pp.path -> string
-  end = struct
-    module String_set = Set.Make (String)
-
-    type 'a test_case = {
-      path : Pp.path;
-      speed_level : speed_level;
-      fn : 'a rrun;
-    }
-
-    type 'a t = {
-      tests : 'a test_case list;
-      (* caches computed from the library values. *)
-      filepaths : String_set.t;
-      doc : (Pp.path, string) Hashtbl.t;
-    }
-
-    let empty () =
-      let tests = [] in
-      let filepaths = String_set.empty in
-      let doc = Hashtbl.create 0 in
-      { tests; filepaths; doc }
-
-    let check_path_is_unique t path =
-      let exn_of_path (`Path (name, _)) =
-        Registration_error (Fmt.strf "Duplicate test name: %s" name)
-      in
-      if String_set.mem (file_of_path path) t.filepaths then
-        raise (exn_of_path path)
-
-    let add t (path, doc, speed_level, fn) =
-      check_path_is_unique t path;
-      let tests = { path; speed_level; fn } :: t.tests in
-      let filepaths = String_set.add (file_of_path path) t.filepaths in
-      Hashtbl.add t.doc path doc;
-      { t with tests; filepaths }
-
-    let tests t = List.rev t.tests
-
-    let doc_of_path t path = try Hashtbl.find t.doc path with Not_found -> ""
-  end
-
   (* global state *)
   type 'a t = {
     (* library values. *)
-    name : string;
     suite : 'a Suite.t;
     (* runtime state. *)
-    mutable errors : string list;
+    mutable errors : unit Fmt.t list;
     (* runtime options. *)
     max_label : int;
     speed_level : speed_level;
@@ -172,10 +117,9 @@ struct
     run_id : string;
   }
 
-  let empty () =
-    let name = Filename.basename Sys.argv.(0) in
+  let empty ~suite_name =
     let errors = [] in
-    let suite = Suite.empty () in
+    let suite = Suite.v ~name:suite_name in
     let max_label = 0 in
     let verbose = false in
     let compact = false in
@@ -186,7 +130,6 @@ struct
     let log_dir = P.getcwd () in
     let run_id = Uuidm.to_string ~upper:true Uuidm.nil in
     {
-      name;
       errors;
       suite;
       max_label;
@@ -245,22 +188,34 @@ struct
       in
       String.concat ~sep:"\n" display_lines ^ "\n"
 
-  let log_dir t = Filename.concat t.log_dir t.run_id
+  let log_dir ~via_symlink t =
+    let via_symlink =
+      (* We don't create symlinks on Windows. *)
+      via_symlink && not Sys.win32
+    in
+    Filename.concat t.log_dir
+      (if via_symlink then Suite.name t.suite else t.run_id)
 
-  let pp_suite_results t =
-    Pp.suite_results ~verbose:t.verbose ~show_errors:t.show_errors ~json:t.json
-      ~compact:t.compact ~log_dir:(log_dir t)
+  let pp_suite_results ({ verbose; show_errors; json; compact; _ } as t) =
+    let log_dir = log_dir ~via_symlink:true t |> maybe_collapse_home in
+    Pp.suite_results ~verbose ~show_errors ~json ~compact ~log_dir
 
-  let pp_event t =
+  let pp_event ~prior_error ~tests_so_far t =
+    let selector_on_failure =
+      (not prior_error) && not (t.verbose || t.show_errors)
+    in
     if not t.json then
       Pp.event ~compact:t.compact ~max_label:t.max_label
-        ~doc_of_path:(Suite.doc_of_path t.suite)
+        ~doc_of_test_name:(Suite.doc_of_test_name t.suite)
+        ~selector_on_failure ~tests_so_far
     else Fmt.nop
 
   let pp_info t =
-    Pp.info ~max_label:t.max_label ~doc_of_path:(Suite.doc_of_path t.suite)
+    Pp.info ~max_label:t.max_label
+      ~doc_of_test_name:(Suite.doc_of_test_name t.suite)
 
-  let output_file t path = Filename.concat (log_dir t) (file_of_path path)
+  let output_file t tname =
+    Filename.concat (log_dir ~via_symlink:true t) (Test_name.file tname)
 
   let color c ppf fmt = Fmt.(styled c string) ppf fmt
 
@@ -268,86 +223,102 @@ struct
 
   let red ppf fmt = Fmt.kstrf (fun str -> red_s ppf str) fmt
 
-  let bold_s fmt = color `Bold fmt
-
-  let pp_error ~verbose ~doc_of_path ~output_file ~tail_errors ppf (path, error)
-      =
-    let logs =
+  let pp_error ~verbose ~doc_of_test_name ~output_file ~tail_errors ~max_label
+      ppf e =
+    let path, error_fmt =
+      match e with `Error (p, f) -> (p, f) | `Exn (p, _, f) -> (p, f)
+    in
+    let pp_logs ppf () =
       let filename = output_file path in
-      if verbose || not (Sys.file_exists filename) then Fmt.strf "%s\n" error
+      if verbose || not (Sys.file_exists filename) then
+        Fmt.pf ppf "%a@," error_fmt ()
       else
         let file = open_in filename in
         let output = read_tail tail_errors file in
         close_in file;
-        Fmt.strf "in `%s`:\n%s" filename output
+        Fmt.pf ppf "%s@,Logs saved to %a.@," output
+          Fmt.(Pp.quoted (styled `Cyan string))
+          (maybe_collapse_home filename)
     in
-    Fmt.pf ppf "-- %s [%s] Failed --\n%s"
-      (short_string_of_path path)
-      (doc_of_path path) logs
+    Fmt.(
+      Pp.with_surrounding_box
+        (const (Pp.event_line ~max_label ~doc_of_test_name) (`Result (path, e)))
+      ++ pp_logs
+      ++ Pp.horizontal_rule
+      ++ cut)
+      ppf ()
 
-  let failure : Pp.run_result -> bool = function
-    | `Ok | `Skip -> false
-    | `Error _ | `Exn _ | `Todo _ -> true
-
-  let has_run : Pp.run_result -> bool = function
+  let has_run : Run_result.t -> bool = function
     | `Ok | `Error _ | `Exn _ -> true
     | `Skip | `Todo _ -> false
 
   let bt () = match Printexc.get_backtrace () with "" -> "" | s -> "\n" ^ s
 
-  let exn path name err =
-    let err = Printf.sprintf "%s%s" err (bt ()) in
-    `Exn (path, name, err)
+  let exn path name pp = `Exn (path, name, Fmt.(pp ++ const lines (bt ())))
 
-  let protect_test path (f : 'a run) : 'a rrun =
+  let protect_test path (f : 'a run) : 'a -> Run_result.t M.t =
    fun args ->
     M.catch
       (fun () -> f args >|= fun () -> `Ok)
-      (function
-        | Check_error err ->
-            let err = Printf.sprintf "Test error: %s%s" err (bt ()) in
-            M.return @@ `Error (path, err)
-        | Failure f -> M.return @@ exn path "failure" f
-        | Invalid_argument f -> M.return @@ exn path "invalid" f
-        | e -> M.return @@ exn path "exception" (Printexc.to_string e))
+      ( (function
+          | Check_error err ->
+              let err = Fmt.(err ++ const string (bt ())) in
+              `Error (path, err)
+          | Failure s -> exn path "failure" Fmt.(const string s)
+          | Invalid_argument s -> exn path "invalid" Fmt.(const string s)
+          | e -> exn path "exception" Fmt.(const exn e))
+      >> M.return )
 
-  let perform_test t args suite =
+  type running_state = { tests_so_far : int; prior_error : bool }
+  (** State that is kept during the test executions. *)
+
+  let perform_test t args { tests_so_far; prior_error } suite =
     let open Suite in
     let test = suite.fn in
-    let pp_event = pp_event t in
+    let pp_event = pp_event t ~prior_error ~tests_so_far in
     M.return () >>= fun () ->
-    pp_event Fmt.stdout (`Start suite.path);
+    pp_event Fmt.stdout (`Start suite.name);
     Fmt.(flush stdout) () (* Show event before any test stderr *);
     test args >|= fun result ->
     (* Store errors *)
-    let () =
+    let errored : bool =
       let pp_error =
         pp_error ~verbose:t.verbose
-          ~doc_of_path:(Suite.doc_of_path t.suite)
+          ~doc_of_test_name:(Suite.doc_of_test_name t.suite)
           ~output_file:(output_file t) ~tail_errors:t.tail_errors
+          ~max_label:t.max_label
       in
-      let error =
+      let error, errored =
         match result with
-        | `Exn (p, n, _) -> [ Fmt.strf "%a" pp_error (p, n) ]
-        | `Error e -> [ Fmt.strf "%a" pp_error e ]
-        | _ -> []
+        | (`Error _ | `Exn (_, _, _)) as e -> ([ Fmt.const pp_error e ], true)
+        | _ -> ([], false)
       in
-      t.errors <- error @ t.errors
+      t.errors <- error @ t.errors;
+      errored
     in
     (* Show any remaining test output before the event *)
     Fmt.(flush stdout ());
     Fmt.(flush stderr ());
-    pp_event Fmt.stdout (`Result (suite.path, result));
-    result
+    pp_event Fmt.stdout (`Result (suite.name, result));
+    let state =
+      { tests_so_far = tests_so_far + 1; prior_error = errored || prior_error }
+    in
+    (state, result)
 
-  let perform_tests t tests args = M.List.map_s (perform_test t args) tests
+  let perform_tests t tests args =
+    M.List.fold_map_s (perform_test t args)
+      { tests_so_far = 0; prior_error = false }
+      tests
 
   let skip_fun _ = M.return `Skip
 
   let skip_label test_case = Suite.{ test_case with fn = skip_fun }
 
   let filter_test_case (regexp, cases) test_case =
-    let (`Path (n, i)) = test_case.Suite.path in
+    let n, i =
+      let tn = test_case.Suite.name in
+      Test_name.(name tn, index tn)
+    in
     let regexp_match = function None -> true | Some r -> Re.execp r n in
     let case_match = function None -> true | Some set -> IntSet.mem i set in
     regexp_match regexp && case_match cases
@@ -360,7 +331,7 @@ struct
            else None)
 
   let redirect_test_output t test_case =
-    let output_file = output_file t test_case.Suite.path in
+    let output_file = output_file t test_case.Suite.name in
     let fn args =
       P.with_redirect output_file (fun () ->
           test_case.fn args >|= fun result ->
@@ -369,13 +340,14 @@ struct
     in
     { test_case with fn }
 
-  let select_speed speed_level (test_case : 'a Suite.test_case) :
-      'a Suite.test_case =
+  let select_speed speed_level (test_case : 'a Suite.test_case as 'tc) : 'tc =
     if compare_speed_level test_case.speed_level speed_level >= 0 then test_case
     else Suite.{ test_case with fn = skip_fun }
 
   let result t test args =
-    P.prepare ~base:t.log_dir ~dir:(log_dir t) ~name:t.name;
+    P.prepare ~base:t.log_dir
+      ~dir:(log_dir ~via_symlink:false t)
+      ~name:(Suite.name t.suite);
     let start_time = P.time () in
     let test =
       if t.verbose then test else List.map (redirect_test_output t) test
@@ -384,53 +356,33 @@ struct
     perform_tests t test args >|= fun results ->
     let time = P.time () -. start_time in
     let success = List.length (List.filter has_run results) in
-    let failures = List.length (List.filter failure results) in
+    let failures = List.length (List.filter Run_result.is_failure results) in
     Pp.{ time; success; failures; errors = List.rev t.errors }
 
   let list_registered_tests t () =
     Suite.tests t.suite
-    |> List.map (fun t -> t.Suite.path)
-    |> List.sort compare_path
+    |> List.map (fun t -> t.Suite.name)
+    |> List.sort Test_name.compare
     |> Fmt.(list ~sep:(const string "\n") (pp_info t) stdout)
 
-  let get_codepoint buf u =
-    Buffer.add_string buf (Printf.sprintf "U+%04X" (Uchar.to_int u))
-
-  let normalize_name (name : string) =
-    let buf = Buffer.create (String.length name * 2) in
-    let get_normalized_char _ _ u =
-      match u with
-      | `Uchar u ->
-          if Uchar.is_char u then
-            match Uchar.to_char u with
-            | ('A' .. 'Z' | 'a' .. 'z' | '0' .. '9' | '_' | '-' | ' ') as c ->
-                Buffer.add_char buf c
-            | _ -> get_codepoint buf u
-          else get_codepoint buf u
-      | `Malformed _ -> Uutf.Buffer.add_utf_8 buf Uutf.u_rep
-    in
-    Uutf.String.fold_utf_8 get_normalized_char () name;
-    Buffer.contents buf
-
-  let register t (name, (ts : 'a test_case list)) =
+  let register (type a) (t : a t) (name, (ts : a test_case list)) : a t =
     let max_label = max t.max_label (String.length name) in
     let test_details =
       List.mapi
-        (fun i (doc, speed, test) ->
-          let path = `Path (name, i) in
+        (fun index (doc, speed, test) ->
+          let path = Test_name.v ~name ~index in
           let doc =
             if doc = "" || doc.[String.length doc - 1] = '.' then doc
             else doc ^ "."
           in
-          (path, doc, speed, protect_test path test))
+          let test a = protect_test path test a in
+          (path, doc, speed, test))
         ts
     in
     let suite = List.fold_left Suite.add t.suite test_details in
     { t with suite; max_label }
 
-  let register_all t tl =
-    let normalize (n, ts) = (normalize_name n, ts) in
-    List.map normalize tl |> List.fold_left register t
+  let register_all t cases = List.fold_left register t cases
 
   let run_tests ?filter t () args =
     let suite = Suite.tests t.suite in
@@ -447,11 +399,11 @@ struct
           let tests = filter_test_cases ~subst:true labels suite in
           result t tests args )
     >|= fun result ->
-    Fmt.(pf stdout) "%a" (pp_suite_results t) result;
+    (pp_suite_results t) Fmt.stdout result;
     result.failures
 
-  let list_tests (tl : 'a test list) =
-    let t = register_all (empty ()) tl in
+  let list_tests (type a) (tl : a test list) =
+    let t = register_all (empty ~suite_name:"") tl in
     list_registered_tests t ();
     M.return ()
 
@@ -473,16 +425,15 @@ struct
 
   let run_with_args ?(and_exit = true) ?(verbose = false) ?(compact = false)
       ?(tail_errors = `Unlimited) ?(quick_only = false) ?(show_errors = false)
-      ?(json = false) ?filter ?(log_dir = default_log_dir ()) name args
-      (tl : 'a test list) =
+      ?(json = false) ?filter ?(log_dir = default_log_dir ()) name (type a)
+      (args : a) (tl : a test list) =
     let speed_level = if quick_only then `Quick else `Slow in
     let random_state = Random.State.make_self_init () in
     let run_id = Uuidm.v4_gen random_state () |> Uuidm.to_string ~upper:true in
     let t =
       {
-        (empty ()) with
+        (empty ~suite_name:name) with
         run_id;
-        name;
         verbose;
         compact;
         tail_errors;
@@ -495,8 +446,12 @@ struct
     let t = register_all t tl in
     ( (* Only print inside the concurrency monad *)
       M.return () >>= fun () ->
-      Fmt.(pf stdout) "Testing %a.\n" bold_s name;
-      Fmt.(pf stdout) "This run has ID `%s`.\n" run_id;
+      let open Fmt in
+      pr "Testing %a.@," (Pp.quoted Fmt.(styled `Bold Suite.pp_name)) t.suite;
+      pr "@[<v>%a@]"
+        (styled `Faint (fun ppf () ->
+             pf ppf "This run has ID %a.@,@," (Pp.quoted string) run_id))
+        ();
       run_tests ?filter t () args )
     >|= fun test_failures ->
     match (test_failures, and_exit) with